+++ src/main/java/com/moulberry/flashback/keyframe/impl/TimelapseKeyframe.java
@@ -77,7 +77,7 @@ public class TimelapseKeyframe extends Keyframe {
     }
 
     @Override
-    public KeyframeChange createHermiteInterpolatedChange(Map<Integer, Keyframe> keyframes, float amount) {
+    public KeyframeChange createHermiteInterpolatedChange(Map<Float, Keyframe> keyframes, float amount) {
         throw new UnsupportedOperationException();
     }
